+++ src/Security/OpenIdConfigurationProviderManager.php
@@ -26,6 +26,11 @@ class OpenIdConfigurationProviderManager
      *         leeway?: int,
      *         cache_duration?: int,
      *         allow_http?: bool,
+     *         http_client_options?: array{
+     *             timeout?: float,
+     *             proxy?: string,
+     *             verify?: bool,
+     *         },
      *     }>,
      * } $config
      */
@@ -81,6 +86,10 @@ class OpenIdConfigurationProviderManager
 
             if (isset($options['allow_http'])) {
                 $providerOptions['allowHttp'] = $options['allow_http'];
+            }
+
+            if (!empty($options['http_client_options'])) {
+                $providerOptions += $options['http_client_options'];
             }
 
             $this->providers[$key] = new OpenIdConfigurationProvider($providerOptions);